refactor(ProtectedRoute): rewrite ProtectedRoute to tsx
