--- src/components/ProtectedRoute.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Navigate } from 'react-router-dom';
-
-const ProtectedRoute = ({ component: Component, ...props }) => {
-  return (
-    props.loggedIn ? <Component {...props} /> : <Navigate to={props.link} replace />
-  );
-};
-
-export default ProtectedRoute;
