+++ src/games/ItGame/index.tsx
