+++ src/games/ItGame/pages/MainPage/MainPage.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from 'react';
-import { useCategories } from '../../ItGame';
 import Header from '../../components/Header/Header';
 import PlayingField from '../../components/PlayingField/PlayingField';
 import InfoSection from '../../components/InfoSection/InfoSection';
+import { useCategories } from '../..';
 
 const MainPage: FC = () => {
   const { categoriesList } = useCategories();
